// stacktags/schema-types.ts
/* eslint-disable */
/**
 * This file was automatically generated by pulumi-provider-scripts.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source Pulumi Schema file,
 * and run "npm run gen-types" to regenerate this file. */
import * as pulumi from "@pulumi/pulumi";
export type ConstructComponent<T extends pulumi.ComponentResource = pulumi.ComponentResource> = (name: string, inputs: any, options: pulumi.ComponentResourceOptions) => T;
export type ResourceConstructor = {
    readonly "stacktags:index:StackTags": ConstructComponent<StackTags>;
};
export type Functions = {};
export abstract class StackTags<TData = any> extends (pulumi.ComponentResource)<TData> {
    constructor(name: string, args: pulumi.Inputs, opts: pulumi.ComponentResourceOptions = {}) {
        super("stacktags:index:StackTags", name, opts.urn ? {} : { name, args, opts }, opts);
    }
}
export interface StackTagsArgs {
    readonly organization?: pulumi.Input<string>;
    readonly project?: pulumi.Input<string>;
    readonly stack?: pulumi.Input<string>;
    readonly configNamespace?: pulumi.Input<string>;
}